moved scheduler out of master
